requirements.txt
altair==4.2.2
attrs==22.2.0
blinker==1.5
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
decorator==5.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
entrypoints==0.4
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.30
idna==3.4
importlib-metadata==6.0.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
langcodes==3.3.0
lazypredict==0.2.12
lightgbm==3.3.5
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.6.3
mdurl==0.1.2
murmurhash==1.0.9
nltk==3.8.1
numpy==1.24.1
packaging==23.0
pandas==1.5.3
pathy==0.10.1
patsy==0.5.3
Pillow==9.4.0
plotly==5.13.0
plotly-express==0.4.1
preshed==3.0.8
protobuf==3.20.3
pyarrow==11.0.0
pydantic==1.10.4
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.10.31
requests==2.28.2
rich==13.3.1
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.12.2
semver==2.13.0
six==1.16.0
sklearn==0.0.post1
smart-open==6.3.0
smmap==5.0.0
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.5
statsmodels==0.13.5
streamlit==1.17.0
tenacity==8.1.0
thinc==8.1.7
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
wasabi==1.1.1
watchdog==2.2.1
wordcloud==1.8.2.2
xgboost==1.7.3
zipp==3.12.0
